requirements.txt
torch
opencv-python
Pillow
